Remove scrabble point value item from TODO
So long as the solver focuses on removing as many values from the
dictionary as possible for a given attempt, it should be working
optimally. This is possible because the dictionary now effectively
serves as a whitelist of possible answers, which I believe are all
equally as likely as each other.